% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weight.R
\name{apply_weight}
\alias{apply_weight}
\title{Applying expert weight}
\usage{
apply_weight()
}
\value{
Writes \code{watershed_prioritization_level5.csv} and
\code{watershed_prioritization_level6.csv}
}
\description{
Applying expert weight
}
