% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weights.R
\name{apply_weights}
\alias{apply_weights}
\title{Applying expert weight to watershed prioritization data}
\usage{
apply_weights(ws_data)
}
\arguments{
\item{ws_data}{A data frame with the variable of interest
(see \code{\link[=generate_canada_dataset]{generate_canada_dataset()}} and \code{\link[=generate_lake_erie_dataset]{generate_lake_erie_dataset()}}).}
}
\value{
A data frame with the same data as \code{ws_data} plus four columns for the
scores computed and four additional four columns including the scaled
score within FEOW if \code{ws_data} includes \code{FEOW_ID}.
}
\description{
Applying expert weight to watershed prioritization data
}
\details{
Use the median values of expert weights used to generate weighted-scores for
four management strategies (weighting schema):
\itemize{
\item Area-based protection
\item Restoration
\item SAR management
\item Invasive species management
}
}
