% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_priority.R
\name{compute_fish_priority_index}
\alias{compute_fish_priority_index}
\title{Calculate rarity indices}
\usage{
compute_fish_priority_index()
}
\description{
Calculate rarity indices
}
\details{
Calculation based on I and Q indices in Minns (1987).
}
\references{
\itemize{
\item Minns, C. K. 'A Method of Ranking Species and Sites for Conservation Using
Presence-Absence Data and Its Application to Native Freshwater Fish in New
Zealand. New Zealand Journal of Zoology 14, no. 1 (January 1987): 43–49.
https://doi.org/10.1080/03014223.1987.10422680.
}
}
