% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_minns.R
\name{compute_minns_Q_I}
\alias{compute_minns_Q_I}
\title{Compute I and Q indices in Minns (1987).}
\usage{
compute_minns_Q_I(pa)
}
\arguments{
\item{pa}{\verb{[matrix]}\cr a presence/absence matrix (site x species).}
}
\description{
Compute I and Q indices in Minns (1987).
}
\references{
\itemize{
\item Minns, C. K. 'A Method of Ranking Species and Sites for Conservation Using
Presence-Absence Data and Its Application to Native Freshwater Fish in New
Zealand. New Zealand Journal of Zoology 14, no. 1 (January 1987): 43–49.
https://doi.org/10.1080/03014223.1987.10422680.
}
}
