% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feow_scaling.R
\name{feow_scaling}
\alias{feow_scaling}
\title{FEOW scaling}
\usage{
feow_scaling()
}
\description{
Scaling ranks level 6 hydrobasins within canadian FEOWs (Freshwater
Ecoregions of the World).
}
\details{
The scaling enables comparison between regions, considering that regions
have varying numbers of watersheds and, consequently, different value ranges
for ranking.
}
