% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_dataset.R
\name{generate_canada_dataset}
\alias{generate_canada_dataset}
\title{Watershed priorization dataset for Canada}
\usage{
generate_canada_dataset()
}
\value{
A data frame with the following columns:
\itemize{
\item HYBAS_ID: Hydrobassin identifier ,
\item FEOW_ID: Freshwater Ecoregions of the World identifier,
\item WSI_n: Watershed stress index from Hirsh-Pearson (2022),
\item FBCI_n: Fish biodiversity change index. Calculate Jaccard dissimilarity
for native spp pool vs. current native + non-native spp pool.
\item CCI_n: Climate change velocities based on CMIP6 SSP2-4.5 forward
velocities for 2040-2070 period.
\item SARI_n: Species at risk management
\item Fish_richness_n: Species richness.
\item Priority_n: Index of priority based on Minns rarity index (1987).
Bote that columns ending with \verb{_n} are normalized.
\item Protected_area_n: Percentage of protected area overlap (capped at 100\% if
multiple types of protected areas overlap within the watershed, resulting in
a total exceeding this limit).
}
}
\description{
Generate a watershed prioritization dataset for the Lake Erie watershed.
The dataset includes seven normalized variables used to prioritize watershed
conservation.
}
\details{
Note that raw data are embeded in the R package in \code{extdata} folder.
The scaling (variables are normalized) enables comparison between regions,
considering that regions have varying numbers of watersheds and,
consequently, different value ranges for ranking.
}
\references{
\itemize{
\item Hirsh-Pearson, Kristen, Chris J. Johnson, Richard Schuster, Roger D.
Wheate, and Oscar Venter. 'Canada’s Human Footprint Reveals Large Intact
Areas Juxtaposed against Areas under Immense Anthropogenic Pressure.' Edited
by Raymond Bradley. FACETS 7 (January 1, 2022): 398–419.
https://doi.org/10.1139/facets-2021-0063.
\item Minns, C. K. 'A Method of Ranking Species and Sites for Conservation Using
Presence-Absence Data and Its Application to Native Freshwater Fish in New
Zealand. New Zealand Journal of Zoology 14, no. 1 (January 1987): 43–49.
https://doi.org/10.1080/03014223.1987.10422680.
}
}
