% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_results.R
\name{generate_canada_results_set}
\alias{generate_canada_results_set}
\title{Generates the set of figures for Canada}
\usage{
generate_canada_results_set(can_data)
}
\arguments{
\item{can_data}{Canada dataset see \code{\link[=generate_canada_dataset]{generate_canada_dataset()}}.}
}
\description{
Generates the set of figures for Canada
}
