% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_erie_dataset.R
\name{generate_lake_erie_dataset}
\alias{generate_lake_erie_dataset}
\title{Watershed prioritization data set for Lake Erie}
\usage{
generate_lake_erie_dataset()
}
\value{
A data frame with the following columns:
\itemize{
\item HYBAS_ID: Hydrobassin identifier ,
\item WSI_n: Watershed stress index from Hirsh-Pearson (2022),
\item FBCI_n: Fish biodiversity change index. Calculate Jaccard dissimilarity
for native spp pool vs. current native + non-native spp pool.
\item CCI_n: Climate change velocities based on CMIP6 SSP2-4.5 forward
velocities for 2040-2070 period.
\item SARI_n: Species at risk management
\item Fish_richness_n: Species richness.
\item Priority_n: Index of priority based on Minns rarity index (1987).
Bote that columns ending with \verb{_n} are normalized.
}
}
\description{
Generate a watershed prioritization dataset for the Lake Erie watershed.
The dataset includes six normalized variables used to prioritize watershed
conservation.
}
\details{
Note that raw data are embedded in the R package in \code{extdata} folder.
The scaling (variables are normalized) enables comparison between regions,
considering that regions have varying numbers of watersheds and,
consequently, different value ranges for ranking. For Lake Erie, only
82 units with sufficient data have been retained.
}
\references{
\itemize{
\item Hirsh-Pearson, Kristen, Chris J. Johnson, Richard Schuster, Roger D.
Wheate, and Oscar Venter. 'Canada’s Human Footprint Reveals Large Intact
Areas Juxtaposed against Areas under Immense Anthropogenic Pressure.' Edited
by Raymond Bradley. FACETS 7 (January 1, 2022): 398–419.
https://doi.org/10.1139/facets-2021-0063.
\item Minns, C. K. 'A Method of Ranking Species and Sites for Conservation Using
Presence-Absence Data and Its Application to Native Freshwater Fish in New
Zealand. New Zealand Journal of Zoology 14, no. 1 (January 1987): 43–49.
https://doi.org/10.1080/03014223.1987.10422680.
}
}
