% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{plot_comparison}
\alias{plot_comparison}
\alias{plot_comparison_pca}
\title{Compare Spatial overlap of high priority watersheds for 2 variables}
\usage{
plot_comparison(
  map,
  feow,
  var1,
  var2,
  xlab = "Priority for area-based protection",
  ylab = "Priority for habitat restoration",
  filename = "protection_vs_restoration.png"
)

plot_comparison_pca(map)
}
\arguments{
\item{map}{\verb{[sf]}\cr  Map that include the geometries of the watersheds
along with the prioritization input variables (normalized).}

\item{feow}{\verb{[sf]}\cr Geometries of the Freshwater Ecoregions of the World (FEOW).}

\item{var1, var2}{\verb{[var]}\cr Variables to be compared.}

\item{xlab, ylab}{\verb{[character string]}\cr axis labels.}

\item{filename}{\verb{[character string]}\cr Name of the output file (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}).}
}
\value{
The function plots the figure and returns \code{TRUE} invisibly.
}
\description{
Compare Spatial overlap of high priority watersheds for 2 variables
}
\details{
Used for:
\itemize{
\item Spatial overlap of high priority watersheds for habitat restoration and
area-based protection;
\item Species at risk management and invasive species management.
}
}
\section{Functions}{
\itemize{
\item \code{plot_comparison_pca()}: PCA to vizualize correlations among scaled variales.

}}
\examples{
\dontrun{
feow <- path_input_data("FEOW_CAN_Extent/FEOW__CAN_Extent.shp") |>
    sf::read_sf()
map <- generate_canada_dataset() |>
    apply_weights() |>
    spatialize_results() |>
    plot_comparison(Prot_rank_feow_scaled, Rest_rank_feow_scaled, feow, Prot_rank_feow_scaled)
}

}
