% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input_variables.R
\name{plot_input_variables}
\alias{plot_input_variables}
\title{Plot input variables}
\usage{
plot_input_variables(map, filename = "normalized_index_values.png")
}
\arguments{
\item{map}{\verb{[sf]}\cr  Map that includes the geometries of the watersheds
along with the prioritization input variables (normalized).}

\item{filename}{\verb{[character string]}\cr Name of the output file (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}).}
}
\value{
The function plots the figure and returns \code{TRUE} invisibly.
}
\description{
Plot the six input variables used to compute the the priority score of
watersheds.
}
\examples{
\dontrun{
generate_canada_dataset() |>
    apply_weights() |>
    spatialize_results() |>
    plot_input_variables()
}
}
