% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_protected_areas.R
\name{plot_protected_areas}
\alias{plot_protected_areas}
\title{Plot percentage of protected area per watershed}
\usage{
plot_protected_areas(map, filename = "protected_areas.png")
}
\arguments{
\item{map}{\verb{[sf]}\cr  Map that include the geometries of the watersheds
along with the prioritization input variables (normalized).}

\item{filename}{\verb{[character string]}\cr Name of the output file (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}).}
}
\value{
The function plots the figure and returns \code{TRUE} invisibly.
}
\description{
Plot percentage of protected area per watershed
}
\examples{
\dontrun{
map <- generate_canada_dataset() |>
    apply_weights() |>
    spatialize_results() |>
    plot_protected_areas(map)
}

}
