% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scale_dependency.R
\name{plot_scale_dependency}
\alias{plot_scale_dependency}
\title{Generates the set of figures for Canada}
\usage{
plot_scale_dependency(map, map5, filename = "scale_dependency.png")
}
\arguments{
\item{map}{\verb{[sf]}\cr Results for Canada see \code{\link[=generate_canada_dataset]{generate_canada_dataset()}}.}

\item{map5}{\verb{[sf]}\cr Results for Canada at level 5.}

\item{filename}{\verb{[character string]}\cr Name of the output file (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}).}
}
\description{
Generates the set of figures for Canada
}
\examples{
\dontrun{
generate_canada_dataset() |>
    apply_weights() |>
    spatialize_results() |>
    plot_scale_dependency(map5)
}
}
