% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\alias{plot_scores_feow}
\title{Plot scores}
\usage{
plot_scores(map, filename = "national_priorities.png")

plot_scores_feow(map, filename = "priorities_by_feow.png")
}
\arguments{
\item{map}{\verb{[sf]}\cr  Map that include the geometries of the watersheds
along with the prioritization input variables (normalized).}

\item{filename}{\verb{[character string]}\cr Name of the output file (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}).

The function plots the figure and returns \code{TRUE} invisibly.}
}
\description{
Plot the four scores based on the six input variables used to compute the
the priority score of watersheds.
}
\section{Functions}{
\itemize{
\item \code{plot_scores_feow()}: Plot scores within FEOW.

}}
\examples{
\dontrun{
map <- generate_canada_dataset() |>
    apply_weights() |>
    spatialize_results() |>
plot_scores(map)
plot_scores_feow(map)
}

}
