% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize_results.R
\name{spatialize_results}
\alias{spatialize_results}
\title{Generates a sf object including results.}
\usage{
spatialize_results(x)
}
\arguments{
\item{x}{\verb{[data.frame]}\cr A data frame including the results.}
}
\description{
Generates a sf object including results.
}
\details{
Use the number of rows to determine whether the geometry of Lake Erie or the
entire Canada should be used.
}
