% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weight.R
\name{apply_weight}
\alias{apply_weight}
\title{Applying expert weight}
\usage{
apply_weight()
}
\value{
write \code{watershed_prioritization_level6.csv}
}
\description{
Median values of expert weights used to generate weighted-scores for:
\itemize{
\item Area-based protection
\item Restoration
\item SAR management
\item Invasive species management
}
}
