% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_results.R
\name{generate_canada_results_set}
\alias{generate_canada_results_set}
\title{Generates the set of figures for Canada}
\usage{
generate_canada_results_set(map_can)
}
\arguments{
\item{map_can}{\verb{[sf]}\cr Results for Canada see \code{\link[=generate_canada_dataset]{generate_canada_dataset()}}.}
}
\description{
Generates the set of figures for Canada
}
\examples{
\dontrun{
 generate_canada_dataset() |>
     apply_weights() |>
     spatialize_results() |>
     generate_canada_results_set()
}
}
