% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Pipeline}
\usage{
run_pipeline(write = TRUE)
}
\arguments{
\item{write}{a logical, should results be written (2 geopackages)?}
}
\description{
This function runs all the steps the reproduce the analysis.
}
